[Feature][Cleanup] Unify flashinfer utils into package structure
Consolidate the previously scattered FlashInfer utilities into a single
unified package at `vllm/utils/flashinfer/`.

Changes:
- Create `vllm/utils/flashinfer/` package with:
  - `core.py`: General FlashInfer compatibility (from utils/flashinfer.py)
  - `moe.py`: MoE-specific utilities (from quantization/utils/flashinfer_utils.py)
  - `__init__.py`: Re-exports all symbols for backwards compatibility
- Convert old `flashinfer_utils.py` to deprecation shim with warning
- All existing imports continue to work unchanged

Benefits:
- Single import path: `from vllm.utils.flashinfer import ...`
- Clear separation: core vs MoE functionality
- Room for future growth (new submodules)
- Matches vLLM conventions for namespaced utils

Fixes #31414
